requirements.txt
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
mysql-connector-python==9.1.0
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
